% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsePLIER.R
\name{commonRows}
\alias{commonRows}
\title{returns the row names in common}
\usage{
commonRows(exp1, exp2)
}
\arguments{
\item{exp1}{One matrix with gene rownames}

\item{exp2}{Another matrix with gene rownames}
}
\description{
returns the row names in common
}
\keyword{internal}
