% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsePLIER.R
\name{crossVal}
\alias{crossVal}
\title{crossVal}
\usage{
crossVal(plierResults, exp, pathwaylist, pathwaylistcv)
}
\arguments{
\item{pathwaylist}{the real prior info matrix}

\item{pathwaylistcv}{the zeroed-out prior info matrix used for PLIER computations}
}
\description{
crossVal
}
\keyword{internal}
