% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreprocessing.R
\name{.filterFeatures}
\alias{.filterFeatures}
\title{Title Filter out genes with constant expression values}
\usage{
.filterFeatures(expr)
}
\arguments{
\item{expr}{Gene expression matrix}
}
\value{
expr
}
\description{
This is an R function that filters out genes with constant expression values from an expression matrix.
Genes with standard deviation less than a threshold value of 1e-10 are considered to have constant expression values
and are removed from the expression matrix.
}
\keyword{internal}
