% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsessGSEA.R
\name{.filterFeaturesgsea}
\alias{.filterFeaturesgsea}
\title{Title  filter out genes with constant expression values}
\usage{
.filterFeaturesgsea(expr)
}
\arguments{
\item{expr}{the gene exp}
}
\description{
Title  filter out genes with constant expression values
}
