% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureRankingFunction.R
\name{FeatureImportance}
\alias{FeatureImportance}
\title{Title feature importance ranking obtained by machine learning.}
\usage{
FeatureImportance(iES_mat, label, type = "rf")
}
\arguments{
\item{iES_mat}{the iES_mat obtained by pathway activation evaluation algorithms}

\item{label}{normal(0) and tumor (1) in patient samples.}

\item{type}{Machine learning algorithms, including Random Forest, Linear Regression, Logistic Regression, k-Nearest Neighbors, etc., default random forest.}
}
\value{
The feature importance ranking
}
\description{
This function uses machine learning algorithm to get the feature importance ranking based on caret package.
}
\examples{
data(GSVAResults)
data(KICHcli)
PathwayImportance = FeatureImportance(GSVAResults, label = KICHcli[3,] )
}
