% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureRankingFunction.R
\name{FeatureImportanceKM}
\alias{FeatureImportanceKM}
\title{Title Significance ranking of survival differences between disturbed and normal-like  samples}
\usage{
FeatureImportanceKM(iES_mat, cli, samplenum = 10)
}
\arguments{
\item{iES_mat}{the iES_mat obtained by pathway activation evaluation algorithms}

\item{cli}{the survival information for patient samples.}

\item{samplenum}{minimum number of normal-like  and perturbed samples}
}
\value{
The survival difference(p-value) between normal-likeand and disturbed samples
}
\description{
This function uses survival analysis to statistically analyze the survival status of disturbed and normal-like  samples in different pathways
}
\examples{
data(GSVAResults)
data(KICHcli)
PathwayImportance = FeatureImportanceKM(GSVAResults, KICHcli )
}
