% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunction.R
\name{GetARIscore}
\alias{GetARIscore}
\title{Title  Obtain adjusted rand index (ARI) score}
\usage{
GetARIscore(iES_mat, label, samplenum = 20, n_iter = 50)
}
\arguments{
\item{iES_mat}{the iES_mat obtained by pathway activation evaluation algorithms.}

\item{label}{normal(0) and tumor (1) in patient samples.}

\item{samplenum}{the number of tumor and normal samples}

\item{n_iter}{number of repeated experiments}
}
\value{
ARI_vec
}
\description{
Title  Obtain adjusted rand index (ARI) score
}
\examples{
data(GSVAResults)
data(KICHcli)
GSVAARIscore = GetARIscore(GSVAResults,label = KICHcli[3,] )
}
