% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsePLIER.R
\name{getAUC}
\alias{getAUC}
\title{getAUC}
\usage{
getAUC(plierResults, exp, pathwaylist)
}
\arguments{
\item{plierResults}{current PLIER result}

\item{exp}{the input exp}

\item{pathwaylist}{the  prior info matrix}
}
\description{
getAUC
}
\keyword{internal}
