% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsePLIER.R
\name{getCutoff}
\alias{getCutoff}
\title{get the p-value cutoff for a specific FDR}
\usage{
getCutoff(plierResults, fdr.cutoff = 0.01)
}
\arguments{
\item{plierResults}{A PLIER result}

\item{fdr.cutoff}{The cross-validation significance cutoff for a pathway.}
}
\description{
get the p-value cutoff for a specific FDR
}
\keyword{internal}
