% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultAnalyze.R
\name{GetReproducibilityScore}
\alias{GetReproducibilityScore}
\title{Title Method evaluation: Reproducibility score}
\usage{
GetReproducibilityScore(exp, iES_mat)
}
\arguments{
\item{iES_mat}{The iES_mat obtained by pathway activation evaluation algorithms.}

\item{exp_feature}{The gene expression matrix.}
}
\value{
The reproducibility score.
}
\description{
Title Method evaluation: Reproducibility score
}
\examples{
data(KICH)
data(KEGGgenesetsID)
KICHfilter <- NormalizeExp(KICH)
pathway_ret <- UseGSVA(KICHfilter, KEGGgenesetsID)
ReproducibilityScore <- GetReproducibilityScore(KICHfilter, pathway_ret)
}
