% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultAnalyze.R
\name{GetROCAUCscore}
\alias{GetROCAUCscore}
\title{Title  Method evaluation: ROC AUC}
\usage{
GetROCAUCscore(sample_feature, sample_labels)
}
\arguments{
\item{sample_feature}{A data frame. The row subscript is the index of the sample. The column subscript is the index of each feature.}

\item{sample_labels}{A data frame. The row subscript is the index of the sample. Only one column represents the label.}
}
\value{
ROC AUC score.
}
\description{
Title  Method evaluation: ROC AUC
}
\examples{
data(KICH)
data(KICHcli)
data(KEGGgenesetsID)
KICHfilter <- NormalizeExp(KICH)
pathway_ret <- UseCORGs(KICHfilter, KEGGgenesetsID)
ROCAUCscore <- GetROCAUCscore(as.data.frame(t(pathway_ret)), as.data.frame(t(KICHcli[3,])))
}
