% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultAnalyze.R
\name{GetRunningTimeScore}
\alias{GetRunningTimeScore}
\title{Title Method evaluation: Running time}
\usage{
GetRunningTimeScore(func = NA, ...)
}
\arguments{
\item{func}{The function that wants to evaluate running time.}

\item{...}{The param list of function.}
}
\value{
The running time score of the method.
}
\description{
Title Method evaluation: Running time
}
\examples{
data(KICH)
data(KEGGgenesetsID)
KICHfilter <- NormalizeExp(KICH)
RunningTimeScore <- GetRunningTimeScore(UseGSVA, KICHfilter, KEGGgenesetsID)
}
