% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsePLIER.R
\name{mapPathway}
\alias{mapPathway}
\title{Rename pathway matrix gene names.}
\usage{
mapPathway(pathway, map)
}
\arguments{
\item{pathway}{the pathway.}

\item{map}{Gene name map.}
}
\description{
Rename pathway matrix gene names.
}
\keyword{internal}
