% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreprocessing.R
\name{NormalizeExp}
\alias{NormalizeExp}
\title{Title Expression matrix standardization}
\usage{
NormalizeExp(exp, method = "none")
}
\arguments{
\item{exp}{Gene expression matrix}

\item{method}{Z-score, quantile, TPM, TMM, min-max}
}
\value{
The normalized matrix
}
\description{
This function is used to standardize the gene expression matrix, which mainly consists of three steps:
Step 1: Filter out the same gene Step 2: Filter out genes with constant expression values Step 3: Normalize the matrix
}
\examples{
data(KICH)
KICHfilter = NormalizeExp(KICH)

}
