% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsePLIER.R
\name{num_pc}
\alias{num_pc}
\title{estimates the number of 'significant' principle components for the SVD decomposition: the minimum k for PLIER}
\usage{
num_pc(exp, method = "elbow", B = 20, seed = NULL)
}
\arguments{
\item{exp}{the same exp as to be used for PLIER (z-score recommended)}

\item{method}{Either "eblow" (fast) or "permutation" (slower, but less heuristic)}

\item{B}{number of permutations}

\item{seed}{seed for reproducibility}
}
\description{
estimates the number of 'significant' principle components for the SVD decomposition: the minimum k for PLIER
}
\keyword{internal}
