% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsePLIER.R
\name{plierResToMarkers}
\alias{plierResToMarkers}
\title{Creates a binary cell-type marker matrix using prior results.}
\usage{
plierResToMarkers(plierResults, pathwaylist, num = 20, index = NULL)
}
\arguments{
\item{plierResults}{A PLIER result}

\item{pathwaylist}{the binary prior information matrix that was used to compute the plierResult.}

\item{num}{The number of marker genes.}

\item{index}{The indecies of PLIER latent variables that are believed to represent cell-type proportions.}
}
\description{
Creates a binary cell-type marker matrix using prior results.
}
\keyword{internal}
