% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunction.R
\name{PlotDensityfall}
\alias{PlotDensityfall}
\title{Title Plot densityfall}
\usage{
PlotDensityfall(iES_mat, path_name, label, title = TRUE)
}
\arguments{
\item{iES_mat}{the iES_mat obtained by pathway activation evaluation algorithms}

\item{path_name}{the pathway name}

\item{label}{normal(0) and tumor (1) in patient samples.}

\item{title}{boolean true or false for including the title (path_name) in the densityfall plot.}
}
\value{
p
}
\description{
Title Plot densityfall
}
\examples{
data(GSVAResults)
data(KICHcli)
PlotDensityfall(GSVAResults, path_name = "HALLMARK_PI3K_AKT_MTOR_SIGNALING" , label = KICHcli[3,] )
}
