% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunction.R
\name{plotggviolin}
\alias{plotggviolin}
\title{Title Plot ggviolin}
\usage{
plotggviolin(ARIscores, compare = F)
}
\arguments{
\item{ARIscores}{a two-column matrix, the first column(methods) and the second column(ARI value)}

\item{compare}{Boolean value indicating whether statistical tests are performed for comparison between groups.}
}
\value{
p
}
\description{
Title Plot ggviolin
}
\examples{
data(ARIscores)
plotggviolin(ARIscores)
}
