% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunction.R
\name{PlotHeatmap}
\alias{PlotHeatmap}
\title{Title Plot Heatmap}
\usage{
PlotHeatmap(iES_mat, label, condition = TRUE, title = NA)
}
\arguments{
\item{iES_mat}{the iES_mat obtained by pathway activation evaluation algorithms.}

\item{label}{normal(0) and tumor (1) in patient samples.}

\item{condition}{boolean value indicating whether samples need to be grouped and colored according to different conditions.}

\item{title}{a string representing the title of the heat map.}
}
\value{
p
}
\description{
Title Plot Heatmap
}
\examples{
data(GSVAResults)
data(KICHcli)
PlotHeatmap(GSVAResults, label = KICHcli[3,])
}
