% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunction.R
\name{PlotiESSurv}
\alias{PlotiESSurv}
\title{Title Plot Kaplan-Meier(KM) survival curve}
\usage{
PlotiESSurv(iES_mat, cli, path_name, sample_num = 20, title = TRUE)
}
\arguments{
\item{iES_mat}{the the iES_mat obtained by pathway activation evaluation algorithms.}

\item{cli}{the survival information for patient samples.}

\item{path_name}{the pathway name.}

\item{sample_num}{the minimum number of patients for survival analysis.}

\item{title}{boolean true or false for including the title (path_name) in the KM survival plot.}
}
\value{
p
}
\description{
Title Plot Kaplan-Meier(KM) survival curve
}
\examples{
data(GSVAResults)
data(KICHcli)
PlotiESSurv(GSVAResults, cli = KICHcli, path_name = "HALLMARK_PI3K_AKT_MTOR_SIGNALING")
}
