% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunction.R
\name{PlotRadar}
\alias{PlotRadar}
\title{Title Plot Radar}
\usage{
PlotRadar(performanceScores)
}
\arguments{
\item{performanceScores}{a data frame, rows are all methods, cols are all scores.}
}
\value{
The radar plot.
}
\description{
Title Plot Radar
}
\examples{
data(KICH)
data(KEGGgenesetsID)
data(KICHcli)
KICHfilter <- NormalizeExp(KICH)
GSVARunningTimeScore <- GetRunningTimeScore(UseGSVA, KICHfilter, KEGGgenesetsID)
GSVApathway_ret <- UseGSVA(KICHfilter, KEGGgenesetsID)
GSVAReproducibilityScore <- GetReproducibilityScore(KICHfilter, GSVApathway_ret)
GSVAROCAUCscore <- GetROCAUCscore(as.data.frame(t(GSVApathway_ret)), as.data.frame(t(KICHcli[3,])))
CORGsRunningTimeScore <- GetRunningTimeScore(UseCORGs, KICHfilter, KEGGgenesetsID)
CORGspathway_ret <- UseCORGs(KICHfilter, KEGGgenesetsID)
CORGsReproducibilityScore <- GetReproducibilityScore(KICHfilter, CORGspathway_ret)
CORGsROCAUCscore <- GetROCAUCscore(as.data.frame(t(CORGspathway_ret)), as.data.frame(t(KICHcli[3,])))
ssGSEARunningTimeScore <- GetRunningTimeScore(UsessGSEA, KICHfilter, KEGGgenesetsID)
ssGSEApathway_ret <- UsessGSEA(KICHfilter, KEGGgenesetsID)
ssGSEAReproducibilityScore <- GetReproducibilityScore(KICHfilter, ssGSEApathway_ret)
ssGSEAROCAUCscore <- GetROCAUCscore(as.data.frame(t(ssGSEApathway_ret)), as.data.frame(t(KICHcli[3,])))
PlotRadar(data.frame(t(data.frame(row.names = c("RunningTimeScore", "ReproducibilityScore", "ROCAUCscore"), 
                                 GSVA = c(GSVARunningTimeScore, GSVAReproducibilityScore, GSVAROCAUCscore),
                                 CORGs = c(CORGsRunningTimeScore, CORGsReproducibilityScore, CORGsROCAUCscore),
                                 ssGSEA = c(ssGSEARunningTimeScore, ssGSEAReproducibilityScore, ssGSEAROCAUCscore)))))
}
