% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunction.R
\name{PlotWaterfall}
\alias{PlotWaterfall}
\title{Title  Plot Waterfall}
\usage{
PlotWaterfall(iES_mat, path_name, label, title = TRUE)
}
\arguments{
\item{iES_mat}{the iES_mat obtained by pathway activation evaluation algorithms}

\item{path_name}{the pathway name}

\item{label}{normal(0) and tumor (1) in patient samples.}

\item{title}{boolean true or false for including the title (path_name) in the Waterfall plot.}
}
\value{
p
}
\description{
Title  Plot Waterfall
}
\examples{
data(GSVAResults)
data(KICHcli)
PlotWaterfall(GSVAResults, path_name = "HALLMARK_PI3K_AKT_MTOR_SIGNALING" , label = KICHcli[3,] )
}
