% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UseiPath.R
\name{remove_data}
\alias{remove_data}
\title{remove genes with 0 sd}
\usage{
remove_data(exp)
}
\arguments{
\item{exp}{is the gene expression matrix.}
}
\value{
a processed matrix
}
\description{
This function uses remove non-informative genes.
}
\keyword{internal}
