% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsePLIER.R
\name{rowNorm}
\alias{rowNorm}
\title{z-score each row of the exp}
\usage{
rowNorm(x)
}
\arguments{
\item{x}{gene expression matrix, with genes in rows}
}
\description{
z-score each row of the exp
}
\keyword{internal}
