% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UseiPath.R
\name{UseiPath}
\alias{UseiPath}
\title{iES calculation Function}
\usage{
UseiPath(exp, pathwaylist, BPPARAM = NULL, nPro = 0)
}
\arguments{
\item{exp}{the gene expression matrix.}

\item{pathwaylist}{the pathway genesets.}

\item{BPPARAM}{parameters from the BiocParallel.}

\item{nPro}{number of processors (default = 0).}
}
\value{
a matrix with rows corresponding to the pathways and columns corresponding to the patients.
}
\description{
This function calculates the iES matrix which is the core of iPath.
}
\examples{
data(KICH)
data(KEGGgenesetsID)
KICHfilter <- NormalizeExp(KICH)
iPathResults = UseiPath(KICHfilter, KEGGgenesetsID)
}
\keyword{calculatioin.}
\keyword{iES}
\keyword{statistics}
