% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Usepathifier.R
\name{Usepathifier}
\alias{Usepathifier}
\title{Title pathifier,a pathway activation evaluation algorithm.}
\usage{
Usepathifier(
  exp,
  paths,
  label = NULL,
  ranks = NULL,
  attempts = 100,
  maximize_stability = TRUE,
  logfile = "",
  samplings = NULL,
  min_exp = 4,
  min_std = 0.4
)
}
\arguments{
\item{exp}{gene expression matrix, where n is the number of genes and m the number of samples}

\item{paths}{a list of p pathways}

\item{label}{normal(0) and tumor (1) in patient samples.}

\item{ranks}{external knowledge on the ranking of the m samples, if exists (to use initial guess)}

\item{attempts}{number of repeats  to stability}

\item{maximize_stability}{throw away components leading to low stability of sampling noise.}

\item{logfile}{name of the file the log should be written to (use stdout if empty).}

\item{samplings}{a matrix specifying the samples that should be chosen in each sampling attempt, chooses a random matrix if samplings is NULL.}

\item{min_exp}{the minimal expression considered as a real signal. Any values below are thresholded to be min_exp.}

\item{min_std}{the minimal allowed standard deviation of each gene.}
}
\value{
the iES_mat obtained by pathifier
}
\description{
Title pathifier,a pathway activation evaluation algorithm.
}
\examples{
data(KICH)
data(KEGGgenesetsID)
KICHfilter <- NormalizeExp(KICH)
pathifierResults = Usepathifier(KICHfilter, KEGGgenesetsID)
}
