% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsePLAGE.R
\name{UsePLAGE}
\alias{UsePLAGE}
\title{Title  Pathway activation evaluation algorithm：PLAGE}
\usage{
UsePLAGE(
  exp,
  pathwaylist,
  kernelMethods = c("Gaussian", "Poisson", "none"),
  abs.ranking = FALSE,
  min.sz = 1,
  max.sz = Inf,
  parallel.sz = 1L,
  mx.diff = TRUE,
  tau = NA,
  ssgsea.norm = TRUE,
  verbose = TRUE,
  BPPARAM = SerialParam(progressbar = verbose)
)
}
\arguments{
\item{exp}{the gene expression matrix.}

\item{pathwaylist}{the pathway genesets.}

\item{kernelMethods}{character string denoting the kernel to use during the non-parametric estimation of the cumulative distribution function of expression levels across samples
when method="gsva". By default, kernelMethods="Gaussian" which is suitable when input expression values are continuous, such as microarray.
When input expression values are integer counts, such as RNA-seq experiments should be set to kernelMethods="Poisson"}

\item{abs.ranking}{When abs.ranking=FALSE (default) a modified Kuiper statistic is used to calculate enrichment scores.}

\item{min.sz}{minimum size of the resulting gene sets.}

\item{max.sz}{maximum size of the resulting gene sets.}

\item{parallel.sz}{number of threads of execution to use when doing the calculations in parallel.}

\item{mx.diff}{two approaches to calculate the enrichment statistic (ES) from the KS random walk statistic. mx.diff=FALSE: ES is calculated as the maximum
distance of the random walk from 0. mx.diff=TRUE (default): ES is calculated as the magnitude difference between the largest positive and negative random walk deviations.}

\item{tau}{exponent defining the weight of the tail in the random walk}

\item{ssgsea.norm}{Logical, set to TRUE (default) . When ssgsea.norm=FALSE this last normalization     step is skipped.}

\item{verbose}{Gives information about each calculation step. Default: FALSE.}

\item{BPPARAM}{An object of class BiocParallelParam specifiying parameters.}
}
\value{
IES
}
\description{
Title  Pathway activation evaluation algorithm：PLAGE
}
\examples{
data(KICH)
data(KEGGgenesetsID)
KICHfilter <- NormalizeExp(KICH)
PLAGEResults = UsePLAGE(KICHfilter, KEGGgenesetsID)
}
