% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UseSEPA.R
\name{UseSEPA}
\alias{UseSEPA}
\title{Title Pathway activation evaluation algorithm：SEPA}
\usage{
UseSEPA(exp, pathwaySparseAdj)
}
\arguments{
\item{exp}{the gene expression matrix.}

\item{pathwaySparseAdj}{the pathway genesets.}
}
\value{
a matrix with rows corresponding to the pathways and columns corresponding to the patients.
}
\description{
Title Pathway activation evaluation algorithm：SEPA
}
\examples{
data(KICH)
data(KEGGgenenetworkID)
KICHfilter <- NormalizeExp(KICH)
SEPAResults <- UseSEPA(KICHfilter, KEGGgenenetworkID)
}
